refactor: profileReducer to profileReducer and postsReducer
