+++ src/Components/Main/Pages/MessagesPage/Dialogs/DialogsContainer.js
@@ -1,8 +1,9 @@
 import { Dialogs } from './Dialogs';
 import { useSelector } from 'react-redux';
+import { dialogsSelector } from '../../../../../redux/selectors/messages';
 
 export const DialogsContainer = () => {
-  const dialogs = useSelector((state) => state.messages.dialogs);
+  const dialogs = useSelector(dialogsSelector);
 
   return (
     <Dialogs dialogs={dialogs} />
