+++ src/redux/actions/posts.js
@@ -1,0 +1,3 @@
+import { ADD_POST } from '../types/posts';
+
+export const addPost = (text) => ({type: ADD_POST, payload: text});
