+++ src/redux/reducers/profile.js
@@ -1,13 +1,6 @@
-import { SET_PROFILE_SUCCESS, SET_STATUS_SUCCESS, ADD_POST } from '../types/profile';
+import { SET_PROFILE_SUCCESS, SET_STATUS_SUCCESS } from '../types/profile';
 
 const initialState = {
-  posts: [
-    {id: 1, postText: 'Hi, Max'},
-    {id: 2, postText: 'Hi, Eugene'},
-    {id: 3, postText: 'Hi, Yuri'},
-    {id: 4, postText: 'Hi, Alexei'},
-    {id: 5, postText: 'Hi, Andrey'}
-  ],
   profile: {},
   status: '',
 };
@@ -23,16 +16,6 @@ export const profileReducer = (state = initialState, action) => {
       return {
         ...state,
         status: action.payload
-      };
-    case ADD_POST:
-      let lastPostId = state.posts[state.posts.length - 1].id;
-      const newPost = {
-        id: ++lastPostId,
-        postText: action.payload
-      };
-      return {
-        ...state,
-        posts: [...state.posts, newPost]
       };
     default:
       return state;
