+++ src/redux/reducers/reducers.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import { profileReducer } from './profile';
 import { messagesReducer } from './messages';
+import { postsReducer } from './posts';
 import { usersReducer } from './users';
 import { authReducer } from './auth';
 import { loadingReducer } from './loading';
@@ -8,6 +9,7 @@ import { loadingReducer } from './loading';
 export const rootReducer = combineReducers({
   profile: profileReducer,
   messages: messagesReducer,
+  posts: postsReducer,
   users: usersReducer,
   auth: authReducer,
   loading: loadingReducer
