+++ src/redux/selectors/messages.js
@@ -1,0 +1,2 @@
+export const dialogsSelector = (state) => state.messages.dialogs;
+export const messagesSelector = (state) => state.messages.messages;
