+++ src/redux/selectors/posts.js
@@ -1,0 +1,1 @@
+export const postsSelector = (state) => state.posts.posts;
