+++ src/redux/selectors/profile.js
@@ -1,3 +1,2 @@
-export const postsSelector = (state) => state.profile.posts;
 export const profileSelector = (state) => state.profile.profile;
 export const statusSelector = (state) => state.profile.status;
