+++ src/redux/types/posts.js
@@ -1,0 +1,1 @@
+export const ADD_POST = 'ADD_POST';
