+++ src/redux/types/profile.js
@@ -2,4 +2,3 @@ export const SET_PROFILE_REQUEST = 'SET_PROFILE_REQUEST';
 export const SET_PROFILE_SUCCESS = 'SET_PROFILE_SUCCESS';
 export const SET_STATUS_REQUEST = 'SET_STATUS_REQUEST';
 export const SET_STATUS_SUCCESS = 'SET_STATUS_SUCCESS';
-export const ADD_POST = 'ADD_POST';
